pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avo"
version = "0.1.0"
description = "AVO: Agentic Variation Operators for Autonomous Evolutionary Search"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
gpu = [
    "torch>=2.0",
]
all = [
    "avo[dev,gpu]",
]

[project.scripts]
avo = "avo.cli:main"

[tool.setuptools.packages.find]
include = ["avo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
